@@ -9,8 +9,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
@@ -22,10 +21,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringRunner;
 
 @ActiveProfiles("test")
-@RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 public class PlanControllerIT {
 
@@ -132,8 +129,8 @@ public class PlanControllerIT {
     // Given Plan Exists
 
     // When Deleting Plan
-    final ResponseEntity<String> response =
-        testRestTemplate.exchange("/plans/5", HttpMethod.DELETE, null, String.class);
+    final ResponseEntity<String> response = testRestTemplate.exchange("/plans/5", HttpMethod.DELETE, null,
+        String.class);
 
     // Then Status Code Is No Content
     assertThat(response.getStatusCode(), is(HttpStatus.NO_CONTENT));
